Adding es6 clarification
Even though I know about these I see them so rarely that I had to pause sometimes and think for a while to realize that it is just plain ES6 syntax. So I guess newcomers would be confused a lot more!
Especially the prewritten parts like the util functions and the user datasource use the defaulting and renaming of destructured variables a lot.